Chapter 10
I made a game menu to which I added the challenges 2 and 3 from the end
of chapter summary. I changed the way the MadLib add opens, instead of
the root frame it opens as a separate window (as do all of the options
on the root menu). I added the GUI version of the Guess My Number Game
which comes from Chapter 3. The Restaurant button displays a sample cake
selection menu which also calculates the price.

I consulted:

http://stackoverflow.com/questions/6920302/passing-argument-in-python-tkinter-button-command
to see how to pass arguments in a command option used with a button.

I used:
http://docs.python.org/2/tutorial/errors.html
to handle ValueErrors from leaving the Qty column empty in the
Restaurant app.